Ryteplan_send_bq_to_fs_gcs/requirements.txt
functions-framework==3.*
google.cloud.bigquery
google.cloud.storage
google.cloud.firestore